+++ R/plot_cat_count_by_year.R
@@ -8,14 +8,22 @@
 #' counts per year will be plotted.
 #' @param mark_cutoff Logical. Whether to show temporal cut-off of previous data.
 #' @param n Integer. Number of most common categories to display. All other categories ignored.
+#' @param scales 	Should scales be fixed ("fixed", the default) or free in the
+#' y dimension ("free_y")?
 #' @return Stacked Bar plot ggplot plot.
 #' @export
 #' @importFrom dplyr select all_of filter mutate
-#' @importFrom ggplot2 ggplot aes geom_bar facet_wrap geom_vline
+#' @importFrom ggplot2 ggplot aes geom_bar facet_wrap geom_vline guides guide_axis
 plot_cat_count_by_year <- function(tbl, date_col, y,
                              mark_cutoff = TRUE,
-                             n = 4L) {
+                             n = 4L,
+                             scales = c("fixed", "free_y")) {
 
+  if (is.integer(tbl[[y]])) {
+    tbl[[y]] <- as.character(tbl[[y]])
+  }
+
+  scales <- rlang::arg_match(scales)
   prev_cutoff <- lubridate::year(
     get_date_ceiling(tbl, date_col, "year")
     )
@@ -36,7 +44,8 @@ plot_cat_count_by_year <- function(tbl, date_col, y,
     geom_bar(
       position = "dodge"
     ) +
-    facet_wrap(~.data[[y]], scales = "free_y")
+    facet_wrap(~.data[[y]], scales = scales) +
+    guides(x = guide_axis(check.overlap = TRUE))
 
   if (mark_cutoff) {
     p <- p +