+++ R/plot_histogram.R
@@ -4,6 +4,10 @@
 #' @param x Character string. Name of variable in table to plot.
 #' @param position Character string.  Bar position.
 #' @param bins Integer. Number of bins.
+#' @param log Logical. If TRUE, log transform x variable with `log(x + 1)`
+#' (to avoid returning `Inf` values) or `log(x + 1 -min(x))` if negative numbers
+#' exist in `x` (to avoid returning `NaN` values).
+#'
 #' @importFrom ggplot2 ggplot aes geom_histogram after_stat ylab facet_grid scale_y_continuous geom_hline
 #' @return Histogram ggplot plot.
 #' @export
@@ -12,8 +16,21 @@ plot_histogram <- function(tbl, x,
                              "dodge",
                              "mirror", "facet"
                            ),
-                           bins = 30L) {
+                           bins = 30L,
+                           log = FALSE) {
   position <- rlang::arg_match(position)
+
+  if (log) {
+    any_neg <- any(tbl[[x]] < 0)
+    if (any_neg) {
+      min <- min(tbl[[x]], na.rm = TRUE)
+      x_label <- glue::glue("log({x} + 1 - min({x}))")
+    } else {
+      min <- 0
+      x_label <- glue::glue("log({x} + 1)")
+    }
+    tbl[[x]] <- log(tbl[[x]] + 1 - min)
+  }
 
   if (position == "dodge") {
     p <- tbl %>%
@@ -33,7 +50,7 @@ plot_histogram <- function(tbl, x,
       )) +
       geom_histogram(position = "stack", bins = bins) +
       scale_y_continuous(labels = abs) +
-      geom_hline(yintercept = 0) +
+      geom_hline(yintercept = 0, linewidth = 0.2) +
       ylab("count")
   }
   if (position == "facet") {
@@ -41,6 +58,9 @@ plot_histogram <- function(tbl, x,
       ggplot(aes(x = .data[[x]], fill = .data[["tbl"]])) +
       geom_histogram(position = "stack", bins = bins) +
       facet_grid(~ .data[["tbl"]])
+  }
+  if (log) {
+    p <- p + xlab(x_label)
   }
   p
 }