+++ R/plot_missing.R
@@ -1,14 +1,23 @@
 #' Create bar plot of missing value (NA) counts across variables in a tbl
 #'
 #' @param tbl Combined table of previous and current data. Output of [combine_tbls()].
-#' @param pk_tbl Tibble of primary key information. Usually stored in userData$pk_tbl.
+#' @param pk_tbl Tibble of primary key information. Usually stored in
+#' `session$userData$pk[[session$userData$pk_tbl_name]]`.
 #' @param exclude Columns to exclude from plot.
-#' @importFrom ggplot2 ggplot aes geom_bar labs ylab
+#' @param compare_pk Logical. Whether to compare table missing values to overall
+#' primary key values.
+#' @importFrom ggplot2 ggplot aes geom_bar labs scale_x_continuous
 #' @return Missing value ggplot bar plot.
 #' @export
-plot_missing <- function(tbl, pk_tbl, exclude = NULL) {
+plot_missing <- function(tbl, pk_tbl, exclude = NULL, compare_pk = FALSE) {
   exclude <- c("tbl_name", exclude)
-  join_pk(tbl, pk_tbl) %>%
+  x_lab <- "% missing"
+  if (compare_pk) {
+    tbl <- join_pk(tbl, pk_tbl)
+    x_lab <- paste(x_lab, "(compared to primary key)")
+  }
+
+  tbl %>%
     dplyr::select(-dplyr::any_of(exclude)) %>%
     dplyr::group_by(tbl) %>%
     naniar::miss_var_summary() %>%
@@ -20,9 +29,10 @@ plot_missing <- function(tbl, pk_tbl, exclude = NULL) {
       )
     ) +
     geom_bar(stat = "identity", position = "dodge") +
+    scale_x_continuous(limits = c(0, 100)) +
     labs(
-      title = "Missingness by Table",
-      x = "Percent Missing",
+      title = "Missingness across Table",
+      x = x_lab,
       y = "Variable"
     )
 }