+++ R/plot_modal_utils.R
@@ -13,6 +13,9 @@ var_type_fn <- function(var_type) {
          character_or_factor = function(x) {
            is.character(x) | is.factor(x)
          },
+         character_or_factor_or_integer = function(x) {
+           is.character(x) | is.factor(x) | is.integer(x)
+         },
          all = function(x) {
            TRUE
          }